rewrote run_changed scripts
