+++ package.json
@@ -38,8 +38,8 @@
     "pretest:coverage": "mkdirp coverage",
     "ci:coverage": "lcov-result-merger 'packages/**/lcov.info' 'lcov-all.info'",
     "test:coverage": "lcov-result-merger 'packages/**/lcov.info' | coveralls",
-    "test:changed": "node scripts/run_changed.js",
-    "test:changed:firestore": "ts-node-script scripts/run_changed_firestore.ts",
+    "test:changed": "ts-node-script scripts/ci-test/run_changed_no_firestore.js",
+    "test:changed:firestore": "ts-node-script scripts/ci-test/run_changed_firestore.ts",
     "test:setup": "node tools/config.js",
     "test:saucelabs": "node scripts/run_saucelabs.js",
     "docgen:js": "node scripts/docgen/generate-docs.js --api js",
